% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin-A function that simulates a binomial experiment}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of trials per iteration}

\item{p}{probability of success in each trial}
}
\value{
barplot of number of successes from all iterations
}
\description{
mybin-A function that simulates a binomial experiment
}
\examples{
mybin(iter=100,n=10,p=0.5)
}
