% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{a data sample}

\item{fun}{the function we would like to estimate}

\item{alpha}{1- (confidence level)}

\item{cx}{size of numbers printed on graph}

\item{...}{any other parameters we might need}
}
\value{
graph of bootstrap distribution and confidence interval
}
\description{
myboot2
}
\examples{
myboot2(iter=10000,x=c(1,2,3),fun="mean")
}
