% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp - a function that uses the CLT and Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of samples}

\item{lambda}{parameter in Poisson dist}

\item{...}{any other parameters for the function}
}
\value{
histograms of the sample mean, barplot of sampled y, and probability function for poisson
}
\description{
mycltp - a function that uses the CLT and Poisson Distribution
}
\examples{
mycltp(n=2,iter=10000,lambda=10)
}
