% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik() -- Uses a graphical method to calculate MLEs}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{log of a function}

\item{x}{vector of data}

\item{param}{range of possible values for parameter}

\item{...}{additional parameters}
}
\value{
a graph of the lfun and list containing index, parameter, max value, and slope
}
\description{
mymaxlik() -- Uses a graphical method to calculate MLEs
}
\examples{
mymaxlik(x=c(9,8,7),param=seq(0,20,length=1000),lfun=function(x,param) log(dexp(x,rate=param)))
}
