% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of normal distribution}

\item{sigma}{standard deviation of normal distribution}

\item{a}{value that you want to find the area to the left of}
}
\value{
a graph of the area under the curve and the area under the curve
}
\description{
myncurve
}
\examples{
myncurve(mu=10,sigma=5,a=6)
}
