% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{myread}
\usage{
myread(csv, dird)
}
\arguments{
\item{csv}{.CSV file name}

\item{dird}{directory to csv files}
}
\value{
imports a csv file into r
}
\description{
myread
}
\examples{
\dontrun{myread("DDT.csv")}
}
