% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets - a function that determines the number of tickets an airline should sell}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of available seats on the flight}

\item{gamma}{probability plane will truly be overbooked}

\item{p}{probability any given passenger will show up}
}
\value{
plots of binomial distribution and normal approximation, and the number of tickets to be sold in each case
}
\description{
ntickets - a function that determines the number of tickets an airline should sell
}
\examples{
ntickets(400,0.02,0.95)
}
